Audit of all types; use modules instead of Namespaces
