+++ src/components/StripeProvider.tsx
@@ -2,7 +2,12 @@ import React, { useEffect } from 'react';
 
 import NativeStripeSdk from '../NativeStripeSdk';
 import { isAndroid, shouldAttributeExpo } from '../helpers';
-import type { AppInfo, InitStripeParams, InitialiseParams } from '../types';
+import type {
+  AppInfo,
+  ThreeDSecure,
+  InitStripeParams,
+  InitialiseParams,
+} from '../types';
 import pjson from '../../package.json';
 
 const EXPO_PARTNER_ID = 'pp_partner_JBN7LkABco2yUu';
@@ -10,9 +15,13 @@ const EXPO_PARTNER_ID = 'pp_partner_JBN7LkABco2yUu';
 /**
  *  Stripe Provider Component Props
  */
-export type Props = InitStripeParams & {
+export interface Props {
+  publishableKey: string;
+  merchantIdentifier?: string;
+  threeDSecureParams?: ThreeDSecure.ConfigurationParams;
+  stripeAccountId?: string;
   children: React.ReactElement | React.ReactElement[];
-};
+}
 
 const repository: any = pjson.repository;
