+++ src/hooks/useStripe.tsx
@@ -1,5 +1,5 @@
 import type {
-  PaymentMethodCreateParams,
+  PaymentMethod,
   ApplePay,
   PaymentSheet,
   CreatePaymentMethodResult,
@@ -15,14 +15,14 @@ import type {
   InitPaymentSheetResult,
   PresentPaymentSheetResult,
   ConfirmPaymentSheetPaymentResult,
-  ConfirmSetupIntent,
+  SetupIntent,
   CreateTokenResult,
   PayWithGooglePayResult,
   GooglePayInitResult,
   GooglePay,
   CreateGooglePayPaymentMethodResult,
   OpenApplePaySetupResult,
-  CreateTokenParams,
+  Token,
   VerifyMicrodepositsParams,
   CollectBankAccountParams,
 } from '../types';
@@ -76,8 +76,8 @@ export function useStripe() {
 
   const _createPaymentMethod = useCallback(
     async (
-      data: PaymentMethodCreateParams.Params,
-      options: PaymentMethodCreateParams.Options = {}
+      data: PaymentMethod.CreateParams,
+      options: PaymentMethod.CreateOptions = {}
     ): Promise<CreatePaymentMethodResult> => {
       return createPaymentMethod(data, options);
     },
@@ -85,7 +85,7 @@ export function useStripe() {
   );
 
   const _createToken = useCallback(
-    async (params: CreateTokenParams): Promise<CreateTokenResult> => {
+    async (params: Token.CreateParams): Promise<CreateTokenResult> => {
       return createToken(params);
     },
     []
@@ -108,8 +108,8 @@ export function useStripe() {
   const _confirmPayment = useCallback(
     async (
       paymentIntentClientSecret: string,
-      data: PaymentMethodCreateParams.Params,
-      options: PaymentMethodCreateParams.Options = {}
+      data: PaymentMethod.ConfirmParams,
+      options: PaymentMethod.ConfirmOptions = {}
     ): Promise<ConfirmPaymentResult> => {
       return confirmPayment(paymentIntentClientSecret, data, options);
     },
@@ -157,8 +157,8 @@ export function useStripe() {
   const _confirmSetupIntent = useCallback(
     async (
       paymentIntentClientSecret: string,
-      data: ConfirmSetupIntent.Params,
-      options: ConfirmSetupIntent.Options = {}
+      data: SetupIntent.ConfirmParams,
+      options: SetupIntent.ConfirmOptions = {}
     ): Promise<ConfirmSetupIntentResult> => {
       return confirmSetupIntent(paymentIntentClientSecret, data, options);
     },
@@ -199,7 +199,7 @@ export function useStripe() {
   );
 
   const _isGooglePaySupported = useCallback(
-    async (params?: GooglePay.IsGooglePaySupportedParams): Promise<boolean> => {
+    async (params?: GooglePay.IsSupportedParams): Promise<boolean> => {
       return isGooglePaySupported(params);
     },
     []
@@ -214,7 +214,7 @@ export function useStripe() {
 
   const _presentGooglePay = useCallback(
     async (
-      params: GooglePay.PresentGooglePayParams
+      params: GooglePay.PresentParams
     ): Promise<PayWithGooglePayResult> => {
       return presentGooglePay(params);
     },
