+++ src/types/SetupIntent.ts
@@ -1,90 +1,77 @@
-import type { Nullable, StripeError } from '.';
-import type { AuBECSDebitFormComponent } from './components/AuBECSDebitForm';
+import type { FormDetails } from './components/AuBECSDebitFormComponent';
+import type { Type } from './PaymentMethod';
+import type { BillingDetails } from './Common';
+import type { LastPaymentError } from './PaymentIntent';
 import type { NextAction } from './NextAction';
-import type {
-  PaymentMethod,
-  PaymentMethods,
-  PaymentMethodCreateParams,
-} from './PaymentMethods';
-
-export interface SetupIntent {
+export interface Result {
   id: string;
   clientSecret: string;
-  lastSetupError: Nullable<ConfirmSetupIntent.LastPaymentError>;
-  created: Nullable<string>;
+  lastSetupError: LastPaymentError | null;
+  created: string | null;
   livemode: boolean;
-  paymentMethodId: Nullable<string>;
-  status: SetupIntents.Status;
-  paymentMethodTypes: PaymentMethods.Types[];
-  usage: SetupIntents.FutureUsage;
-  description: Nullable<string>;
+  paymentMethodId: string | null;
+  status: Status;
+  paymentMethodTypes: Type[];
+  usage: FutureUsage;
+  description: string | null;
   nextAction: NextAction | null;
 }
 
-export namespace ConfirmSetupIntent {
-  export type LastPaymentError = StripeError<string> & {
-    paymentMethod: PaymentMethod;
-  };
+export type Params =
+  | CardParams
+  | IdealParams
+  | BancontactParams
+  | SofortParams
+  | AuBecsDebitParams
+  | SepaParams
+  | PaymentMethodCreateParams.USBankAccountParams;
 
-  export type Params =
-    | CardParams
-    | IdealParams
-    | BancontactParams
-    | SofortParams
-    | AuBecsDebitParams
-    | SepaParams
-    | PaymentMethodCreateParams.USBankAccountParams;
+export interface ConfirmOptions {}
 
-  export interface Options {}
-
-  export interface BaseParams {
-    billingDetails?: PaymentMethods.BillingDetails;
-  }
-
-  export interface CardParams extends BaseParams {
-    type: 'Card';
-  }
-
-  export interface IdealParams extends BaseParams {
-    type: 'Ideal';
-    bankName?: string;
-  }
-
-  export interface SofortParams extends BaseParams {
-    type: 'Sofort';
-    country: string;
-  }
-  export interface BancontactParams extends Required<BaseParams> {
-    type: 'Bancontact';
-  }
-
-  export interface SepaParams extends Required<BaseParams> {
-    type: 'SepaDebit';
-    iban: string;
-  }
-
-  export interface AuBecsDebitParams {
-    type: 'AuBecsDebit';
-    formDetails: AuBECSDebitFormComponent.FormDetails;
-  }
+export interface BaseParams {
+  billingDetails?: BillingDetails;
 }
 
-export namespace SetupIntents {
-  export type FutureUsage =
-    | 'Unknown'
-    | 'None'
-    | 'OnSession'
-    | 'OffSession'
-    | 'OneTime';
+export interface CardParams extends BaseParams {
+  type: 'Card';
+}
 
-  export enum Status {
-    Succeeded = 'Succeeded',
-    RequiresPaymentMethod = 'RequiresPaymentMethod',
-    RequiresConfirmation = 'RequiresConfirmation',
-    Canceled = 'Canceled',
-    Processing = 'Processing',
-    RequiresAction = 'RequiresAction',
-    RequiresCapture = 'RequiresCapture',
-    Unknown = 'Unknown',
-  }
+export interface IdealParams extends BaseParams {
+  type: 'Ideal';
+  bankName?: string;
+}
+
+export interface SofortParams extends BaseParams {
+  type: 'Sofort';
+  country: string;
+}
+export interface BancontactParams extends Required<BaseParams> {
+  type: 'Bancontact';
+}
+
+export interface SepaParams extends Required<BaseParams> {
+  type: 'SepaDebit';
+  iban: string;
+}
+
+export interface AuBecsDebitParams {
+  type: 'AuBecsDebit';
+  formDetails: FormDetails;
+}
+
+export type FutureUsage =
+  | 'Unknown'
+  | 'None'
+  | 'OnSession'
+  | 'OffSession'
+  | 'OneTime';
+
+export enum Status {
+  Succeeded = 'Succeeded',
+  RequiresPaymentMethod = 'RequiresPaymentMethod',
+  RequiresConfirmation = 'RequiresConfirmation',
+  Canceled = 'Canceled',
+  Processing = 'Processing',
+  RequiresAction = 'RequiresAction',
+  Unknown = 'Unknown',
 }
